requirements.txt
gradio
pandas
numpy
python-dotenv
faiss-cpu
huggingface_hub
